validation 추가

+++ src/main/java/com/example/post/service/impl/PostServiceImpl.java
@@ -1,11 +1,11 @@
 package com.example.post.service.impl;
 
+import com.example.comment.repository.CommentRepository;
 import com.example.common.dto.PageList;
 import com.example.common.dto.PageRequest;
 import com.example.common.dto.Result;
 import com.example.common.dto.ResultType;
-import com.example.common.exception.DataNotFoundException;
-import com.example.common.exception.NoAuthorityException;
+import com.example.common.exception.NotAllowedOperationException;
 import com.example.post.model.Post;
 import com.example.post.repository.PostRepository;
 import com.example.post.service.PostService;
@@ -20,6 +20,7 @@ import java.util.List;
 public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
+    private final CommentRepository commentRepository;
 
     @Override
     public PageList<Post> getPostList(PageRequest pageRequest, Post post) {
@@ -38,6 +39,10 @@ public class PostServiceImpl implements PostService {
     @Override
     @Transactional
     public Result deletePost(Post post) {
+        if (commentRepository.getTotalCount(post.getPostNo()) > 0) {
+            throw new NotAllowedOperationException();
+        }
+
         postRepository.deletePost(post.getPostNo());
         return new Result(ResultType.OK);
     }
